Rename main folder to package name
