+++ nukeserversocket/controllers/__init__.py
@@ -1,3 +1,3 @@
 """Main modules that deal with the Nuke code editor widget."""
-from .nuke_script_editor import NukeScriptEditor
 from .nuke_controllers import CodeEditor
+from .nuke_script_editor import NukeScriptEditor
