+++ nukeserversocket/controllers/nuke_script_editor.py
