+++ nukeserversocket/local/fake_script_editor.py
