+++ nukeserversocket/util.py
