+++ nukeserversocket/widgets/about_widget.py
