+++ nukeserversocket/widgets/connections_widget.py
