+++ nukeserversocket/widgets/error_dialog.py
