+++ nukeserversocket/widgets/log_widgets.py
