+++ nukeserversocket/widgets/settings_widget.py
